Removed option to not split metabolites before ray calculation to simplify postprocessing
Because we can now assume that all external metabolites are either inputs or outputs (since we split them into virtual in- and outputs), we don't have to take care about the ones that could be both anymore.
Also, the unsplit_metabolites method was doing the same as the postprocessing. The postprocessing was computationally much more expensive. This is removed.